Drives First Draft
Some numbers like the large LT drive are set high to test where the ceiling resides, I expect it to be honed in. Pushing so we're on the same version, not because I'm done working on things. Stuff I think we should include with this update:  Rail rework, Kart-KP buff/rework, pskellon hull price, C5 RR nerf, and spellcheck.

Note all ships are going to have their base maneuverability stats recalculated likely tomorrow to be accurate to the vanilla thrust to weight ratios instead of estimations.